removes SampleTable
