+++ ui/src/App.js
@@ -1,13 +1,12 @@
 import React, { Component } from 'react';
 import LandingPage from './components/LandingPage'
 import '../node_modules/patternfly-react/dist/css/patternfly-react.css'
-import SampleTable from './components/landingpage/SampleTable'
 
 class App extends Component {
   render() {
     return (
       <div className="App">
-        <SampleTable />
+        <LandingPage />
       </div>
     );
   }
