+++ ui/src/components/landingpage/Header.js
@@ -9,7 +9,7 @@ class Header extends Component {
   render() {
     return (
       <div className="header">
-        <h1>App Grid Header</h1>
+        <h1>Mobile Apps</h1>
       </div>
     );
   }
